requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
dateparser==1.2.0
numpy==1.24.3
